+++ test/test-applications/integrations/TestApplication.SqlClient.Microsoft/Program.cs
@@ -149,11 +149,8 @@ internal static class Program
 
     private static (string DatabasePassword, string Port) ParseArgs(string[] args)
     {
-        if (args?.Length != 2)
-        {
-            throw new ArgumentException($"{nameof(TestApplication.SqlClient)}: requires two command-line arguments: <dbPassword> <dbPort>");
-        }
-
-        return (DatabasePassword: args[0], Port: args[1]);
+        return (
+            DatabasePassword: ArgumentHelper.GetRequiredArgument(args, "--password"),
+            Port: ArgumentHelper.GetRequiredArgument(args, "--port"));
     }
 }